requirements.txt
fastapi
uvicorn[standard]
requests
pydantic
python-dotenv
google-cloud-aiplatform
langchain
langchain-experimental
langchain-google-vertexai
